[Test] Move archived setting test into its own test class (#116460)
As title says, the archived setting needs to be moved out to not
interfere with other test methods in the same test class.

Resolves: #111798 Resolves: #111777 Resolves: #111774 Resolves: #111799
